Move ILXQtTaskbarAbstractBackend to panel directory
- It is now a global instance
