+++ panel/backends/lxqttaskbartypes.h
